The current updates are not being successfully implemented in the mobile
